bench, parity-tests: move four benches' assertion loops into parity fixtures
raise_bare_class_slot_kinds, getattr_hook_binding, generator_pep479 and
iteration_protocol each carried assertion loops beside the loop their
`.jitstats` sidecars score. In raise_bare_class_slot_kinds those loops held
11248 of the file's 11249 `guard_failures`.

The assertions move to pyre/extra_tests/parity_tests, which carries no
jitstats sidecar and cross-checks against CPython, pyre and pypy:

    raise_bare_class_slot_contracts.py
    stop_iteration_mro_consumers.py
    generator_pep479_subclass_paths.py
    getattr_hook_deopt_arms.py

Re-recorded, as loops_compiled/bridges_compiled/guard_failures:

    raise_bare_class_slot_kinds   9/11/11249 -> 1/0/1
    getattr_hook_binding          10/1/212   -> 1/0/1
    generator_pep479              3/1/201    -> 1/0/1  (loops_aborted 10 -> 0)
    iteration_protocol            5/6/1202   -> 1/2/401

dynasm, cranelift and wasm agree on every counter. `loops_compiled` falls
because the extra loops belonged to the assertion blocks, not to the measured
loop: all four files return to the counters committed at 6aa4b0d3bb2^, before
#1605 folded the assertions in.
